commit#10

+++ HomeWork3/calks/additional/CalculatorWithCounterClassic.java
@@ -3,7 +3,7 @@ package HomeWork3.calks.additional;
 import HomeWork3.calks.simple.CalculatorWithMathExtends;
 
 public class CalculatorWithCounterClassic extends CalculatorWithMathExtends {
-    private int countOperation = 0;
+    private long countOperation = 0;
     public void incrementCountOperation(){       //должен увеличивать внутренний счётчик в калькуляторе
         this.countOperation ++;
     }
